refactor: Remove unused common script src files
