--- common/script/src/content/card-types.js
@@ -1,22 +1,0 @@
-let cardTypes = {
-  greeting: {
-    key: 'greeting',
-    messageOptions: 4,
-    yearRound: true
-  },
-  nye: {
-    key: 'nye',
-    messageOptions: 5
-  },
-  thankyou: {
-    key: 'thankyou',
-    messageOptions: 4,
-    yearRound: true
-  },
-  valentine: {
-    key: 'valentine',
-    messageOptions: 4
-  }
-};
-
-export default cardTypes;
