--- common/script/src/content/health-potion.js
@@ -1,11 +1,0 @@
-import {translator as t} from './helpers';
-
-let potion = {
-  type: 'potion',
-  text: t('potionText'),
-  notes: t('potionNotes'),
-  value: 25,
-  key: 'potion',
-};
-
-export default potion;
