--- common/script/src/content/pets-mounts/index.js
@@ -1,53 +1,0 @@
-import {transform, defaults} from 'lodash';
-
-//--------------------------------------------------
-// Pets and Mounts are generated by eggs * potions
-//
-// { <egg_key>-<potion_key>: true }
-//
-// <egg_key> - the key of the egg the animal hatched from
-// <potion_key> - the potion that hatched the animal
-//
-//--------------------------------------------------
-// Special Pets and Mounts
-// { <animal_type>-<fake_potion>: <transifex_key> }
-//
-// <animal_type> - could be existing animal or a totally new animal
-// <fake_potion> - if a unique animal, potion is Base, if it's a special version of an existing animal, the fake potion is a different from a normal hatching potion
-//--------------------------------------------------
-
-import {
-  drop as dropHatchingPotions,
-  premium as premiumHatchingPotions,
-} from '../hatching-potions';
-import dropEggs from '../eggs/drops';
-import questEggs from '../eggs/quest';
-
-import specialPets from './special-pets';
-import specialMounts from './special-mounts';
-
-let dropPets = generateAnimalSet(dropEggs, dropHatchingPotions);
-let premiumPets = generateAnimalSet(dropEggs, premiumHatchingPotions);
-let questPets = generateAnimalSet(questEggs, dropHatchingPotions);
-let dropMounts = generateAnimalSet(dropEggs, dropHatchingPotions);
-let premiumMounts = generateAnimalSet(dropEggs, premiumHatchingPotions);
-let questMounts = generateAnimalSet(questEggs, dropHatchingPotions);
-
-function generateAnimalSet(eggSet, potionSet) {
-  return transform(eggSet, (m, egg) => {
-    defaults(m, transform(potionSet, (m2, pot) => {
-      return m2[egg.key + "-" + pot.key] = true;
-    }));
-  });
-}
-
-export default {
-  dropPets: dropPets,
-  dropMounts: dropMounts,
-  premiumPets: premiumPets,
-  premiumMounts: premiumMounts,
-  questPets: questPets,
-  questMounts: questMounts,
-  specialPets: specialPets,
-  specialMounts: specialMounts,
-}
