--- common/script/src/content/pets-mounts/special-mounts.js
@@ -1,13 +1,0 @@
-let specialMounts = {
-  'BearCub-Polar':       'polarBear',
-  'LionCub-Ethereal':    'etherealLion',
-  'MantisShrimp-Base':   'mantisShrimp',
-  'Turkey-Base':         'turkey',
-  'Mammoth-Base':        'mammoth',
-  'Orca-Base':           'orca',
-  'Gryphon-RoyalPurple': 'royalPurpleGryphon',
-  'Phoenix-Base':        'phoenix',
-  'JackOLantern-Base':   'jackolantern'
-}
-
-export default specialMounts;
