--- common/script/src/content/pets-mounts/special-pets.js
@@ -1,14 +1,0 @@
-let specialPets = {
-  'Wolf-Veteran':      'veteranWolf',
-  'Wolf-Cerberus':     'cerberusPup',
-  'Dragon-Hydra':      'hydra',
-  'Turkey-Base':       'turkey',
-  'BearCub-Polar':     'polarBearPup',
-  'MantisShrimp-Base': 'mantisShrimp',
-  'JackOLantern-Base': 'jackolantern',
-  'Mammoth-Base':      'mammoth',
-  'Tiger-Veteran':     'veteranTiger',
-  'Phoenix-Base':      'phoenix'
-}
-
-export default specialPets;
