--- common/script/src/content/quests/holiday.js
@@ -1,126 +1,0 @@
-import {each, defaults} from 'lodash';
-import {
-  translator as t,
-  setQuestSetDefaults,
-} from '../helpers';
-
-let holidayQuests = {
-  evilsanta: {
-    text: t('questEvilSantaText'),
-    notes: t('questEvilSantaNotes'),
-    completion: t('questEvilSantaCompletion'),
-    boss: {
-      name: t('questEvilSantaBoss'),
-      hp: 300,
-      str: 1
-    },
-    drop: {
-      items: [
-        {
-          type: 'mounts',
-          key: 'BearCub-Polar',
-          text: t('questEvilSantaDropBearCubPolarMount')
-        }
-      ],
-      gp: 20,
-      exp: 100
-    }
-  },
-  evilsanta2: {
-    text: t('questEvilSanta2Text'),
-    notes: t('questEvilSanta2Notes'),
-    completion: t('questEvilSanta2Completion'),
-    collect: {
-      tracks: {
-        text: t('questEvilSanta2CollectTracks'),
-        count: 20
-      },
-      branches: {
-        text: t('questEvilSanta2CollectBranches'),
-        count: 10
-      }
-    },
-    drop: {
-      items: [
-        {
-          type: 'pets',
-          key: 'BearCub-Polar',
-          text: t('questEvilSanta2DropBearCubPolarPet')
-        }
-      ],
-      gp: 20,
-      exp: 100
-    }
-  },
-  egg: {
-    text: t('questEggHuntText'),
-    notes: t('questEggHuntNotes'),
-    completion: t('questEggHuntCompletion'),
-    value: 1,
-    collect: {
-      plainEgg: {
-        text: t('questEggHuntCollectPlainEgg'),
-        count: 100
-      }
-    },
-    drop: {
-      items: [
-        {
-          type: 'eggs',
-          key: 'Egg',
-          text: t('questEggHuntDropPlainEgg')
-        }, {
-          type: 'eggs',
-          key: 'Egg',
-          text: t('questEggHuntDropPlainEgg')
-        }, {
-          type: 'eggs',
-          key: 'Egg',
-          text: t('questEggHuntDropPlainEgg')
-        }, {
-          type: 'eggs',
-          key: 'Egg',
-          text: t('questEggHuntDropPlainEgg')
-        }, {
-          type: 'eggs',
-          key: 'Egg',
-          text: t('questEggHuntDropPlainEgg')
-        }, {
-          type: 'eggs',
-          key: 'Egg',
-          text: t('questEggHuntDropPlainEgg')
-        }, {
-          type: 'eggs',
-          key: 'Egg',
-          text: t('questEggHuntDropPlainEgg')
-        }, {
-          type: 'eggs',
-          key: 'Egg',
-          text: t('questEggHuntDropPlainEgg')
-        }, {
-          type: 'eggs',
-          key: 'Egg',
-          text: t('questEggHuntDropPlainEgg')
-        }, {
-          type: 'eggs',
-          key: 'Egg',
-          text: t('questEggHuntDropPlainEgg')
-        }
-      ],
-      gp: 0,
-      exp: 0
-    }
-  },
-};
-
-let questDefaults = (name) => {
-  return  {
-    completion: t(`quest${name}Completion`),
-    canBuy: () => { return false; },
-    category: 'pet',
-  }
-};
-
-setQuestSetDefaults(holidayQuests, questDefaults);
-
-export default holidayQuests;
