--- common/script/src/content/time-traveler-stable.js
@@ -1,14 +1,0 @@
-import {translator as t} from './helpers';
-
-let stable = {
-  pets: {
-    'Mammoth-Base': t('mammoth'),
-    'MantisShrimp-Base': t('mantisShrimp')
-  },
-  mounts: {
-    'Mammoth-Base': t('mammoth'),
-    'MantisShrimp-Base': t('mantisShrimp')
-  }
-};
-
-export default stable;
