+++ tasks/gulp-eslint.js
@@ -12,7 +12,6 @@ const COMMON_FILES = [
   // @TODO remove these negations as the files are converted over.
   '!./common/script/index.js',
   '!./common/script/content/index.js',
-  '!./common/script/src/**/*.js',
   '!./common/script/public/**/*.js',
 ];
 const TEST_FILES = [
