+++ tasks/gulp-tests.js
@@ -136,7 +136,7 @@ gulp.task('test:content:clean', (cb) => {
 });
 
 gulp.task('test:content:watch', ['test:content:clean'], () => {
-  gulp.watch(['common/script/src/content/**', 'test/**'], ['test:content:clean']);
+  gulp.watch(['common/script/content/**', 'test/**'], ['test:content:clean']);
 });
 
 gulp.task('test:content:safe', ['test:prepare:build'], (cb) => {
@@ -346,7 +346,7 @@ gulp.task('test:all', (done) => {
   'lint',
   'test:e2e:safe',
   'test:common:safe',
-  // 'test:content:safe',
+  'test:content:safe',
   // 'test:server_side:safe',
   'test:karma:safe',
   'test:api-legacy:safe',
